+++ core/src/utilities/helpers.js
@@ -81,6 +81,7 @@ export const prependOrigin = path => {
 };
 
 export const containsAllSegments = (sourceUrl, targetPathSegments) => {
+  console.warn(sourceUrl, targetPathSegments);
   const mandatorySegments = sourceUrl.split('/');
   let result = true;
