+++ core/test/services/routing.spec.js
@@ -4,8 +4,8 @@ const expect = chai.expect;
 const assert = chai.assert;
 const sinon = require('sinon');
 const MockBrowser = require('mock-browser').mocks.MockBrowser;
-const routing = require('../src/services/routing');
-import { deepMerge } from '../src/utilities/helpers.js';
+const routing = require('../../src/services/routing');
+import { deepMerge } from '../../src/utilities/helpers.js';
 import { afterEach } from 'mocha';
 
 describe('Routing', () => {
@@ -671,7 +671,7 @@ describe('Routing', () => {
   });
 
   describe('defaultChildNodes', () => {
-    const routing = rewire('../src/services/routing');
+    const routing = rewire('../../src/services/routing');
     const getDefaultChildNode = routing.__get__('getDefaultChildNode');
     const getPathData = function() {
       return {
